requirements.txt
Flask==3.1.0
Flask-Babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
email-validator==2.2.0
python-dotenv==1.0.1
pytest==8.3.5
